🎨 Move display formatting out of domain
