+++ Project/Widget/Sources/MainIcon+Widget.swift
@@ -1,0 +1,14 @@
+import DomainLayerInterface
+
+extension MainIcon {
+    var fillSystemImage: String {
+        switch self {
+        case .drop:
+            return "drop.fill"
+        case .heart:
+            return "heart.fill"
+        case .cloud:
+            return "cloud.fill"
+        }
+    }
+}
